comps/requirements.txt
aiofiles
aiohttp
docarray
docx2txt
fastapi
httpx
kubernetes
langchain
langchain-community
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-sdk
Pillow
prometheus-fastapi-instrumentator
pypdf
python-multipart
pyyaml
requests
shortuuid
uvicorn
transformers
torch
pytest
accelerate==0.26.0
python-dotenv
motor